server/requirements.txt
fastapi
uvicorn[standard]
py-real-esrgan
Pillow
torch
